pyproject.toml
[project]
name = "fieldtheory"
version = "0.1.0"
description = "The unifying field theory of the GenesisAeon stack: full Lagrangian formulation of S∝A/S∝V duality, medium modulation, cosmic-moment collapse and entropy-table integration."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "GenesisAeon Team", email = "team@genesisaeon.org" }]
keywords = ["field-theory", "lagrangian", "entropy", "sympy", "physics", "genesisaeon"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Physics",
]
dependencies = [
    "sympy>=1.13.0",
    "numpy>=2.0.0",
    "rich>=13.0.0",
    "typer>=0.15.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
stack = [
    "entropy-governance>=0.1.0",
    "entropy-table>=1.0.1",
    "medium-modulation>=0.1.0",
    "cosmic-moment>=0.1.0",
    "implosive-genesis>=0.4.0",
]
dev = ["pytest>=8.0", "pytest-cov>=5.0", "ruff>=0.9"]
docs = ["mkdocs>=1.6", "mkdocstrings[python]>=0.27", "mkdocs-material>=9.6"]

[project.scripts]
ft = "fieldtheory.cli:app"

[project.urls]
Repository = "https://github.com/GenesisAeon/fieldtheory"
Issues = "https://github.com/GenesisAeon/fieldtheory/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/fieldtheory"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "B", "I", "W", "UP", "C4", "SIM"]
ignore = ["B008"]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["S101"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--cov=fieldtheory --cov-report=term-missing -v"

[tool.coverage.run]
source = ["src"]
omit = ["tests/*"]
